requirements.txt
elasticsearch[async]~=8.16.0
pytelegrambotapi~=4.25.0
loguru~=0.7.3
python-dotenv~=1.0.1
fastapi~=0.115.0
pydantic~=2.10.4
requests~=2.32.3
uvicorn~=0.27.1
gunicorn~=22.0.0
python-multipart~=0.0.9